requirements.txt
feedparser
numpy
opencv_python
requests
graia-application-mirai
graia-broadcast